spug_api/requirements.txt
apscheduler==3.6.3
Django==2.2.13
channels==2.3.1
channels_redis==2.4.1
paramiko==2.7.2
django-redis==4.10.0
requests==2.22.0
GitPython==3.0.8
python-ldap==3.2.0
openpyxl==3.0.3
python-gitlab==2.6.0